refactor(UserDataInputContainer): 닉네임, 아이디 입력창 컴포넌트 분리
